requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

aiohttp==3.9.1
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.29.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.1.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
openvino==2025.2.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==10.4.0
psutil==7.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic-settings==2.3.4
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
scipy==1.16.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.176
ultralytics-thop==2.0.15
urllib3==2.5.0
uvicorn==0.30.1
watchfiles==1.1.0
websockets==15.0.1
